refactor: 도메인 기준으로 도메인이 아닌 inboud, outbound를 분리한다 && test: Race::renderRacingSnapshot 실패 테스트
