+++ src/main/java/race/README.md
@@ -8,4 +8,3 @@
    - 고민: 자동차가 자신의 위치를 직접 출력하는게 좋을까? 아니면 렌더링 역할을 담당하는 객체를 만들고 자동차들로 부터 위치를 받아서 출력할까? 도메인은 자동차&경주인데 도메인이 뷰와 엮이는게 과연 좋을까?
  - 사용자 입력을 받아 자동차 수를 세팅
    - 고민: main함수를 만들어야할까? 레이싱과 콘솔(입력)을 분리하는게 좋지않을까..?
- -
