+++ src/main/java/race/Race.java
@@ -1,5 +1,7 @@
 package race;
 
+import race.outbound.RenderRaceSnapshotViaConsole;
+
 import java.util.Collection;
 
 public class Race {
@@ -7,16 +9,23 @@ public class Race {
 
     private final JudgeCarMovement judgeCarMovement;
 
-    public Race(Collection<Car> cars, JudgeCarMovement judgeCarMovement) {
+    private final RenderRaceSnapshot renderRaceSnapshot;
+
+    public Race(Collection<Car> cars, JudgeCarMovement judgeCarMovement, RenderRaceSnapshot renderRaceSnapshot) {
         this.cars = cars;
         this.judgeCarMovement = judgeCarMovement;
+        this.renderRaceSnapshot = renderRaceSnapshot;
     }
 
     public void moveCars() {
-        cars.stream().forEach((car) -> {
+        cars.forEach((car) -> {
             if(judgeCarMovement.judge()) {
                 car.move();
             }
         });
+    }
+
+    public void renderRacingSnapshot() {
+
     }
 }
