+++ src/main/java/race/inboud/GameConsole.java
@@ -1,4 +1,9 @@
-package race;
+package race.inboud;
+
+import race.Car;
+import race.JudgeCarMovement;
+import race.Race;
+import race.outbound.RenderRaceSnapshotViaConsole;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -20,8 +25,7 @@ public class GameConsole {
             cars.add(new Car());
         }
 
-        Race race = new Race(cars, new JudgeCarMovement(new Random()));
-        RenderRaceSnapshot renderRaceSnapshot = new RenderRaceSnapshot();
+        Race race = new Race(cars, new JudgeCarMovement(new Random()), new RenderRaceSnapshotViaConsole());
 
         for (int i = 0; i < numberOfMovement; i++) {
             race.moveCars();
